requirements.txt
streamlit
langchain
langchain-google-genai
langgraph
pydantic
python-dotenv
web3
sqlalchemy
plotly
pandas
qrcode
pillow
